always create an initializer when scaffolding
in the ideal case people will stick to the default values for unittest
and source folders.
PyBuilder is all about plugins though, and even people sticking to the
default values will want to configure their plugins sooner or later. In
that case it is best to have the initializer ready (but empty).