allow denying hardlink capabilities from distutils
This resolves issue #56.
The related stdlib bug is http://bugs.python.org/issue8876 , and once it
is closed and shipped to all python versions that PyBuilder supports,
this can be reverted.
As @markmevans found out, it is not possible to build a bdist when the
underlying filesystem does not support hardlinks, because distutils
assumes that it always does.

Setting the property `distutils_issue8876_workaround_enabled` to True
will deny distutils/setuptools the ability to use os.link and result in
copying files instead of hardlinking them.