Introducing new property for insecure pip install
The property install_dependencies_insecure_install is a list of strings.
Each pip package name in that list will be allowed to be installed
insecurely (may be hosted outside of PyPI, may be unverified).
Note that older versions of pip allow insecure packages by default, and
do not have the required options for a force install.