Make it possible to define distutils console_scripts
Probably we want to make this default instead of filling `scripts`.
But I don't like it right now. It's a hard runtime dependency on
setuptools and I have seen `load_entry_point` behave very strangely on
several systems under various conditions.

This also introduces a `pyb_` script (which is cross-platform and should
work on windows but has a hard runtime dependency on setuptools, as
mentioned).
Somehow resolves issue #94 .
Maybe @aelgru wants to chime in - if we are to support cross platform
libraries and programs, then we will have to use `console_scripts`
instead of `scripts`.