implement using requirements files as dependencies
This resolves #96.
A `Project` exposes two new methods, `depends_on_requirements` and
`build_depends_on_requirements` which can be used to include a
file of type `requirements.txt` in the corresponding dependencies.

Note that runtime requirements will be added to the setup.py script
so that runtime dependencies are pulled in.