Allow scripts named like packages by default.
Due to a clash between the script name and the package name it was
previously impossible to create a script called like a package and have
it work by default. The solution was to add to the configuration that
scripts should be stored in a specific directory while building the
setuptools structure.

This changes now does this by default, calling the script directory
"scripts". The old behaviour can be restored by setting the
"dir_dist_scripts" property to None.