Try really hard to work well with coveralls
Now using a custom coverage object so that it can be configured.
The configuration is used to tell coverage to avoid covering the python
stdlib if possible, and also specifies the source root to cover.

Additionally a .coverage file is written in the coverage root, so that
it can be consumed by coveralls.