checking if posix semaphores work on travis
Previously this would lead to an error since the permissions are not set
correctly. The related travis issue I found is closed and references a
404.
This hack is the last thing we need sudo for. If we can get rid of it
that would allow the use of sudo:false in the travis config, which then
runs on a container instead of a VM.