Adding Flake8 Radon and CC properties
flake8 offers an extra option: –max-complexity, which will emit a warning if the McCabe complexity of a function is higher than the value. By default it’s deactivated
flake8 offers an extra option: –max-complexity, which will emit a warning if the Cyclomatic Complexity of a function is higher than the value. By default it’s deactivated