Exposing Flake8 Code Complexity Property (#468)
* Adding Flake8 Radon and CC properties

flake8 offers an extra option: –max-complexity, which will emit a warning if the McCabe complexity of a function is higher than the value. By default it’s deactivated
flake8 offers an extra option: –max-complexity, which will emit a warning if the Cyclomatic Complexity of a function is higher than the value. By default it’s deactivated

* Radon dependency removed.

Flake8 by default doesn't have --radon-max, the parameter depends on radon library

* Adding Missing Unittesting for McCabe Flake8 Property

* Fixing Assert on Flake8 Unit test